Add option to show solution should be done by hand
Some problems should be done by hand on a sheet of paper, and an image 
should be uploaded to a Markdown cell, rather than using the code to 
solve the problem. This commit adds the option --by-hand which takes a 
list of integers to mark that should be done by hand.