Rework SolutionRemover preprocessor so it can always be used
This change renames the SolnRemoverPreprocessor to SolutionRemover. It
also allows the preprocessor to be used for assignments and solutions by
checking the value of the remove_solution key in the resources
dictionary. This simplifes the converters because only one is needed for
each output type. The remove_solution key must be present in the
resources dictionary. Also change the top-level solution start cell to
be "## solution" (case-insensitive) rather than just "solution"
(case-insensitive). This allows the preprocessor to be used when the
word "solution" appears somewhere useful, e.g., in the problem
statement.